Split app.js phase 4c-4: extract website-input module
In-app tests pass (132/132).
