+++ src/modal-manager.js
@@ -2,7 +2,8 @@
 // Extracted verbatim from app.js.
 import { state } from './state.js';
 import { updateHelperStatusIndicator, updateCleanHostsBtnState } from './settings.js';
-import { closeAllLanguagePickers, isAnyLanguagePickerOpen, resetWebsitesImportMenuPosition } from './app.js';
+import { closeAllLanguagePickers, isAnyLanguagePickerOpen } from './app.js';
+import { resetWebsitesImportMenuPosition } from './website-input.js';
 import { closeAllBlocklistMenus } from './blocklists.js';
 import { closeNowBlockingChipMenus } from './render.js';
 import { closeAllPopovers } from './time-inputs.js';
